requirements.txt
edgartools
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0 
boto3

# pip install "apache-airflow[celery]==3.0.2" --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-3.0.2/constraints-3.9.txt"